+++ src/effects/feats/paladinDefaultAura.js
@@ -1,4 +1,5 @@
 export function paladinDefaultAuraEffect(document) {
+  if (!game.modules.get("ActiveAuras")?.active) return document;
   document.effects.forEach((effect) => {
     if ((effect.name ?? effect.label).includes(" - Constant")) {
       // const distance = document.flags.ddbimporter?.dndbeyond?.levelScale?.fixedValue ?? 10;