+++ src/effects/monsterFeatures/beholderEyeRays.js
@@ -1,18 +1,25 @@
-import { baseFeatEffect } from "../specialFeats.js";
 import DDBMacros from "../DDBMacros.js";
+import { forceItemEffect } from "../effects.js";
 
-export async function beholderEyeRaysEffect(document) {
-  setProperty(document, "system.target", { value: 3, width: null, units: "", type: "creature" });
-  setProperty(document, "system.range", { value: 120, long: null, units: "ft" });
+export async function beholderEyeRaysEffect(document, rayNum = 3, range = 120) {
+  setProperty(document, "system.target", { value: rayNum, width: null, units: "", type: "creature" });
+  setProperty(document, "system.range", { value: range, long: null, units: "ft" });
   setProperty(document, "system.damage", { parts: [], versatile: "", value: "" });
   setProperty(document, "system.activation.type", "action");
+  setProperty(document, "system.actionType", "other");
 
-  let effect = baseFeatEffect(document, document.name);
-  await DDBMacros.setItemMacroFlag(document, "monsterFeature", "beholderEyeRay.js");
-  DDBMacros.setMidiOnUseMacroFlag(document, "monsterFeature", "beholderEyeRay.js", ["postActiveEffects"]);
-  // effect.changes.push(DDBMacros.generateMacroChange({ macroValues: `"${document.name}"`, macroType: "monsterFeature", macroName: "beholderEyeRay.js" }));
-  effect.transfer = false;
+  await DDBMacros.setItemMacroFlag(document, "monsterFeature", "eyeRays.js");
+  DDBMacros.setMidiOnUseMacroFlag(document, "monsterFeature", "eyeRays.js", ["postActiveEffects"]);
 
-  document.effects.push(effect);
+  document.system.save = {
+    dc: null,
+    ability: "",
+    scaling: "spell",
+  };
+
+  setProperty(document, "flags.midiProperties.magiceffect", true);
+
+  document.effects = [];
+  document = forceItemEffect(document);
   return document;
 }