+++ src/effects/monsterFeatures/spellReflection.js
@@ -1,0 +1,19 @@
+import DDBMacros from "../DDBMacros.js";
+import { baseEffect } from "../effects.js";
+
+export async function spellReflectionEffect(document) {
+
+  setProperty(document, "system.activation.type", "special");
+  setProperty(document, "system.actionType", "other");
+
+  await DDBMacros.setItemMacroFlag(document, "monsterFeature", "spellReflection.js");
+
+  let effect = baseEffect(document, document.name, { transfer: true, disabled: false });
+  effect.changes.push(
+    DDBMacros.generateOnUseMacroChange({ macroPass: "isSaveSuccess", macroType: "monsterFeature", macroName: "spellReflection.js", document }),
+    DDBMacros.generateOnUseMacroChange({ macroPass: "isAttacked", macroType: "monsterFeature", macroName: "spellReflection.js", document }),
+  );
+  document.effects.push(effect);
+
+  return document;
+}